Added value setters to top panel. Changed manual tab to status tab

+++ Python/PspController.py
@@ -37,7 +37,7 @@ class Controller():
 
   def getRealCurrent(self):
     with self.processLock:
-      value = float(self.connection.getValue(self.deviceWriteRealCurrent)) * 1000
+      value = int(float(self.connection.getValue(self.deviceWriteRealCurrent)) * 1000)
     return value
 
   def getPreRegulatorVoltage(self):
@@ -52,7 +52,7 @@ class Controller():
 
   def getTargetCurrent(self):
     with self.processLock:
-      value = float(self.connection.getValue(self.deviceWriteTargetCurrent)) * 1000
+      value = int(float(self.connection.getValue(self.deviceWriteTargetCurrent)) * 1000)
     return value
 
   # TODO
